防止大数溢出

+++ leetcode/src/main/java/com/leetcode/tag/must1/four/MySqrt.java
@@ -6,13 +6,14 @@ package com.leetcode.tag.must1.four;
 public class MySqrt {
     class Solution {
         public int mySqrt(int x) {
-            int left = 0;
-            int right = x;
+            // 防止大数溢出
+            long left = 0;
+            long right = x;
 
             while (left <= right) {
-                int mid = left + (right - left) / 2;
+                long mid = left + (right - left) / 2;
                 if (mid * mid == x) {
-                    return mid;
+                    return (int) mid;
                 }
                 if (mid * mid > x) {
                     right = mid - 1;
@@ -21,7 +22,7 @@ public class MySqrt {
                 }
             }
 
-            return right;
+            return (int) right;
         }
     }
 }
